+++ Animations/RQBattleViewController.m
@@ -22,6 +22,8 @@
 #import "RQModelController.h"
 #import "M3CoreDataManager.h"
 #import "RQPassthroughView.h"
+#import <MediaPlayer/MediaPlayer.h>
+#import "AppDelegate_iPhone.h"
 
 @implementation RQBattleViewController
 
@@ -188,6 +190,15 @@
 	[self setupGameLoop];
 	[self startAnimation];
 	
+	// If the user would like us to mute the iPod during battles to hear our own sweet music and the iPod is in fact playing
+	didPauseIPod = NO;
+	if ([[[NSUserDefaults standardUserDefaults] objectForKey:@"RQSoundMuteIPod"] boolValue] && [[MPMusicPlayerController iPodMusicPlayer] playbackState] == MPMusicPlaybackStatePlaying ) {
+		NSLog(@"pauseing iPod");
+		[[MPMusicPlayerController iPodMusicPlayer] pause];
+		[[CDAudioManager sharedManager] setMode:kAMM_FxPlusMusic];
+		didPauseIPod = YES;
+	}
+
     [[SimpleAudioEngine sharedEngine] playBackgroundMusic:@"RQ_Battle_Song.m4a" loop:YES];
 }
 
@@ -491,6 +502,21 @@
 {
     [self stopAnimation];
 	[[SimpleAudioEngine sharedEngine] stopBackgroundMusic];
+	
+	// If the user would like us to pause the iPod during battles to hear our own sweet music
+	// Only play the iPod if we previously paused it
+	if ([[[NSUserDefaults standardUserDefaults] objectForKey:@"RQSoundMuteIPod"] boolValue] && didPauseIPod == YES) {
+		NSLog(@"playing iPod");
+		[[MPMusicPlayerController iPodMusicPlayer] play];
+		[[CDAudioManager sharedManager] setMode:kAMM_FxPlusMusicIfNoOtherAudio];
+		
+		// trying to fix a bug where in the sound effect don't play after the first battle if we are pausing iPod
+		[[CDAudioManager sharedManager] setMute:NO];
+		[[CDAudioManager sharedManager] setEnabled:YES];
+		[[SimpleAudioEngine sharedEngine] setMute:NO];
+		[[SimpleAudioEngine sharedEngine] setEnabled:YES];
+	}
+	
 	[delegate battleViewControllerDidEnd:self];
 }